Finish refactoring windows common helpers into common.go, remove unused bin_extract test file
