+++ frontend/windows/common.go
@@ -12,6 +12,7 @@ import (
 	"github.com/moby/buildkit/client/llb"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
 	"github.com/pkg/errors"
+	"golang.org/x/exp/slices"
 )
 
 const (
@@ -20,6 +21,83 @@ const (
 	buildScriptName = "_build.sh"
 	aptCachePrefix  = "jammy-windowscross"
 )
+
+const gomodsName = "__gomods"
+
+func specToSourcesLLB(worker llb.State, spec *dalec.Spec, sOpt dalec.SourceOpts, opts ...llb.ConstraintsOpt) (map[string]llb.State, error) {
+	out := make(map[string]llb.State, len(spec.Sources))
+	for k, src := range spec.Sources {
+		displayRef, err := src.GetDisplayRef()
+		if err != nil {
+			return nil, err
+		}
+
+		pg := dalec.ProgressGroup("Add spec source: " + k + " " + displayRef)
+		st, err := src.AsState(k, sOpt, append(opts, pg)...)
+		if err != nil {
+			return nil, errors.Wrapf(err, "error creating source state for %q", k)
+		}
+
+		out[k] = st
+	}
+
+	opts = append(opts, dalec.ProgressGroup("Add gomod sources"))
+	st, err := spec.GomodDeps(sOpt, worker, opts...)
+	if err != nil {
+		return nil, errors.Wrap(err, "error adding gomod sources")
+	}
+
+	if st != nil {
+		out[gomodsName] = *st
+	}
+
+	return out, nil
+}
+
+func installBuildDeps(deps []string) llb.StateOption {
+	return func(s llb.State) llb.State {
+		if len(deps) == 0 {
+			return s
+		}
+
+		sorted := slices.Clone(deps)
+		slices.Sort(sorted)
+
+		return s.Run(
+			shArgs("apt-get update && apt-get install -y "+strings.Join(sorted, " ")),
+			dalec.WithMountedAptCache(aptCachePrefix),
+		).Root()
+	}
+}
+
+func withSourcesMounted(dst string, states map[string]llb.State, sources map[string]dalec.Source) llb.RunOption {
+	opts := make([]llb.RunOption, 0, len(states))
+
+	sorted := dalec.SortMapKeys(states)
+	files := []llb.State{}
+
+	for _, k := range sorted {
+		state := states[k]
+
+		// In cases where we have a generated soruce (e.g. gomods) we don't have a [dalec.Source] in the `sources` map.
+		// So we need to check for this.
+		src, ok := sources[k]
+
+		if ok && !dalec.SourceIsDir(src) {
+			files = append(files, state)
+			continue
+		}
+
+		dirDst := filepath.Join(dst, k)
+		opts = append(opts, llb.AddMount(dirDst, state))
+	}
+
+	ordered := make([]llb.RunOption, 1, len(opts)+1)
+	ordered[0] = llb.AddMount(dst, dalec.MergeAtPath(llb.Scratch(), files, "/"))
+	ordered = append(ordered, opts...)
+
+	return dalec.WithRunOptions(ordered...)
+}
 
 func buildBinaries(ctx context.Context, spec *dalec.Spec, worker llb.State, client gwclient.Client, sOpt dalec.SourceOpts, targetKey string) (llb.State, error) {
 	worker = worker.With(installBuildDeps(spec.GetBuildDeps(targetKey)))
