+++ frontend/windows/handle_zip.go
@@ -10,13 +10,6 @@ import (
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
-const (
-	workerImgRef    = "mcr.microsoft.com/mirror/docker/library/ubuntu:jammy"
-	outputDir       = "/tmp/output"
-	buildScriptName = "_build.sh"
-	aptCachePrefix  = "jammy-windowscross"
-)
-
 func handleZip(ctx context.Context, client gwclient.Client) (*gwclient.Result, error) {
 	return frontend.BuildWithPlatform(ctx, client, func(ctx context.Context, client gwclient.Client, platform *ocispecs.Platform, spec *dalec.Spec, targetKey string) (gwclient.Reference, *dalec.DockerImageSpec, error) {
 		sOpt, err := frontend.SourceOptFromClient(ctx, client)
